fix(ci): rename test/ to tests/ to avoid stdlib collision on 3.13
The 'test' package name shadows Python's stdlib test module,
causing ModuleNotFoundError on 3.12+ in CI. Also adds missing
httpx dependency for TestClient.
