+++ .github/workflows/release.yml
@@ -25,10 +25,10 @@ jobs:
         run: |
           pip install --upgrade pip
           pip install -e ".[dev]" 2>/dev/null || pip install -e .
-          pip install pytest pytest-asyncio
+          pip install pytest pytest-asyncio httpx
 
       - name: Run tests
-        run: pytest test/ -v --tb=short
+        run: pytest tests/ -v --tb=short
 
   publish:
     needs: test
